Add button callback
The Button on the main screen is given the onClick attribute, which
invokes a method. This method creates an Intent that acts on a new
class that we have written. A string is attached to the Intent, and
the new class extracts the string and displays it. This replaces what
had been on the screen. It is not exactly clear to me how this happens.

I then added a bit more code just to see if I could have more things
on the new display. I could probably add an onClick method for the
new button to go back to the first display, based on what I have
learned so far.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>