Add Fragment example
This is based on the Fragment tutorial. I tried to adapt it to what
I had been doing, but it was not 100% successful. The Fragment view
only shows the second fragment. I tried a few different things but
the best I could do was show the second fragment twice. I was never
able to get the first fragment to show at all.

Signed-off-by: Patrick Luhmann <patrick.luhmann@gmail.com>